Move C bindings
