+++ gfx/layers/wr/WebrenderColorLayer.cpp
@@ -4,7 +4,7 @@
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
 #include "WebRenderColorLayer.h"
-#include "WebRenderer.h"
+#include "webrender.h"
 
 namespace mozilla {
 namespace layers {
